Add files via upload
This is the sample of mini game. The file will drop from top to bottom and the player need to drag them to the box with correct colour. There is some issues that the circles only move when the mouse move, and when player drag one circle all circles freeze. If anyone know how to improve it please do so. If there is no more problems I will work on those mini games with 7 similar editions for 7 weeks