Add files via upload
Updated the file "Mini Game Week 1.rp" to connect to the main page (the office) and updated the graphics. If you click the pile of papers on the main screen, it will direct you to the mini-game with a screen before the game loads up. Also added a "goal" so that the user will know how many files they will need in order to succeed.